+++ src/kornix/kornixClient.ts
@@ -13,7 +13,6 @@ import type {
   KornixClientOptions,
   KornixCurrentContextDto,
   KornixCurrentIrrigationLayerDto,
-  KornixCurrentPrecipitationLayerDto,
   KornixManualPrecipitationRequestDto,
   KornixManualPrecipitationResponseDto,
   KornixMethodsResponseDto,
@@ -169,10 +168,6 @@ export class KornixClient {
 
   getCurrentIrrigationLayer(seasonYear: number): Promise<KornixCurrentIrrigationLayerDto> {
     return this.request<KornixCurrentIrrigationLayerDto>(kornixEndpoints.currentIrrigationLayer(seasonYear));
-  }
-
-  getCurrentPrecipitationLayer(seasonYear: number): Promise<KornixCurrentPrecipitationLayerDto> {
-    return this.request<KornixCurrentPrecipitationLayerDto>(kornixEndpoints.currentPrecipitationLayer(seasonYear));
   }
 
   getFieldSeasonCatalog(seasonYear: number): Promise<FieldSeasonCatalogDto> {